Mod/#19 loading qr data mod

+++ app/src/main/java/com/example/transparentkey_aos/retrofit2/QRModel.kt
@@ -2,7 +2,7 @@ package com.example.transparentkey_aos.retrofit2
 
 data class QRModel(
     val id: Int,
-    val date: Int
+    val date: Long
 )
 
 data class ResponseBody(
